Adjust CI Tests and update composer.json
The CI Tests have been modified to only trigger on pull requests and pushes to the master branch. Also, the composer.json file has been updated to allow the phpstan/extension-installer plugin alongside the infection/extension-installer. This ensures that our testing workflow is not unnecessarily triggered and aligns our composer settings with the additional plugin.

Signed-off-by: Marcel Strahl <info@marcel-strahl.de>